Fix  Trigonometria

+++ Lab_4/calculate.php
@@ -122,7 +122,7 @@
                                         $arg = calculate($arg);
                                     }
 
-                                    $sum += (int)$arg;
+                                    $sum += (float)$arg;
                                 }
                                 return $sum;
                             }
@@ -139,7 +139,7 @@
                                     if (!is_numeric($args[$i])) {
                                         $args[$i] = calculate($args[$i]);
                                     }
-                                    $minusRez -= (int)$args[$i];
+                                    $minusRez -= (float)$args[$i];
                                 }
                                 return $minusRez;
                             }
@@ -153,7 +153,7 @@
                                     if (!isNum($arg)) {
                                         $arg = calculate($args[$i]);
                                     }
-                                    $sup *= (int)$arg;
+                                    $sup *= (float)$arg;
                                 }
                                 return $sup;
                             }
@@ -171,7 +171,7 @@
                                     if ($args[$i] == 0) {
                                         return "Делить на 0 нельзя";
                                     } else {
-                                        $del /= (int)$args[$i];
+                                        $del /= (float)$args[$i];
                                     }
                                 }
                                 return $del;
